Change function to return bigint
This handles the discrepancy we have between the id types for the jobs table on org/com. The results of these migrations have already been applied to com production, so these migrations cannot be safely applied to com until the type has been changed on com from int -> bigint.

+++ db/main/migrate/20200325130013_add_function_most_recent_job_ids_for_user_repositories_by_states.rb
@@ -2,7 +2,7 @@ class AddFunctionMostRecentJobIdsForUserRepositoriesByStates < ActiveRecord::Mig
   def up
     execute """
     CREATE OR REPLACE FUNCTION
-      most_recent_job_ids_for_user_repositories_by_states(uid int, states varchar default '') RETURNS table (id int) AS
+      most_recent_job_ids_for_user_repositories_by_states(uid int, states varchar default '') RETURNS table (id bigint) AS
       $BODY$
     DECLARE
     rid int;
